backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
